Corrections régressions, bugs divers et variés.
database_servers/cluvfy_samples.md
    Le script database_servers/cluster_verify.sh remplace les samples.

virtualbox/compile_guest_additions.sh -> virtualbox/old/compile_guest_additions.sh
    Le script n'est plus utilisé il est déplacé dans old/

README.md
    Typo

database_servers/clone_master.sh
    Correction regression.

database_servers/install_grid.sh
    Réécriture des fonctions lancant les scripts roots.
    Peut-être que le workaround fonctionne...

database_servers/set_ASM_memory_target_low.sh
    Ajuste également le paramètre memory_target

database_servers/run_all.sh
    N'utilise plus le link db mais le chemin incomplet

dg/add_disks_to.sh
    Revu de code et mise à jour doc wiki.

dg/create_new_dg.sh
    Revu de code.

dg/dg_attr.sql
    Ajout du modeline

disk/remove_unused_partition.sh
    Ce script a été supprimé car il fonctionnait mal et je croyais qu'il
    ne servait pas. Il a été réécrit et est fonctionnel.

dg/drop_oracleasm_disks.sh
    Ne fonctionnait plus à cause de la suppression de
    disk/remove_unused_partition.sh
    Correction bug.

dg/remove_file_for_db.sh
    Typo

dg/show_candidat_disks.sql
    Ajout du modeline

dg/show_dg.sql
    Ajout du modeline

disklib.sh
    Fonction get_unused_disks renomée en
get_unused_disks_without_partitions

disk/create_fs.sh
    Renommage fonction get_unused_disks

disk/create_fs_ocfs2.sh
    Renommage fonction get_unused_disks

disk/create_oracleasm_disks_on_new_disks.sh
    Renommage fonction get_unused_disks
    Le paramètre -ilun a été supprimé, le n° du disque oracleasm est
    déduit de l'existant.
    Si kfod est accessible les disques candidats sont affichés.

global.cfg
    Le serveur K2 se synchronise sur internet par défaut.

memory/root_setup_hp.sh
    Correction aide. (script non fiable encore en test)

memory/show_pages.sh
    Fiabilisation de la lecture de l'alertlog (pages mémoire 4K/2M)
    Pour ASM pas de méthode fiable dans la lecture des pages.

ntp/config_ntp.sh
    Correction de la configuration du fichier chrony.conf

plelib.sh
    La fonction exec_cmd affiche le nom du script ou commande qui plante
    plus lisiblement.

san/add_and_export_lv.sh
    Correction bug dans le test de la variable "export_to"

san/export_lv.sh
    BUG dans le code retour du script.

san/save_targetcli_config.sh
    Réactivation des sauvegardes personnelles.

virtualbox/clone_vm.sh
    Ajustement de la taille du disque du grid pour éviter warning de
    cluvfy.

virtualbox/delete_vm
    Revue de code

virtualbox/restart_vboxdrv.sh
    Revue de code

virtualbox/save_vm
    Revue de code

virtualbox/start_vm
    Réécriture du script.

virtualbox/stop_vm
    Revue de code

virtualbox/get_vm_description
    Supprime les 2 premières lignes de la description.

yum/sync_oracle_repository.sh
    BUG le script plantait s'il n'y avait pas de packages à purger.

+++ yum/sync_oracle_repository.sh
@@ -169,8 +169,15 @@ exec_cmd -c reposync	--newest-only									\
 						--repoid=ol7_latest
 LN
 
-info "Remove old packages :"
-exec_cmd rm $(repomanage --old $infra_olinux_repository_path)
+info "Load packages 2 remove."
+typeset -r packages_2_remove="$(repomanage --old $infra_olinux_repository_path)"
+if [ x"$packages_2_remove" == x ]
+then
+	info "no packages to remove."
+else
+	info "Remove old packages :"
+	exec_cmd rm $(repomanage --old $infra_olinux_repository_path)
+fi
 LN
 
 info "Update local repository :"
